Add chromium_prebuilt.sh for lollipop (Squashed) by Jiangyi (Also includes: omnirom/android_vendor_omni@9a7b0bf)
This script will pull the necessary libs from a complete build and generate proper
makefiles in order to support using a prebuilt Chromium. Files are placed in
prebuilts/chromium/-devicenamehere-, and are not tracked by git, so they are completely
out of the way.

PS3:
-replace git -C with params compatible to old git versions

Change-Id: I3141defcbd4108178881f7203f3105aecc27378b

Prebuilt chromium: update script for lollipop

Change-Id: I3fbef0d23e76a6f5c25579d1ca9336bcfe19290d